refactor validator steps module
